Asynchronously load the data for Worlds.
There may be fallout due to this change, but the synchronous loading of the same data.yml file for successive World Load Events can cause server bootups to take very very long. In our case we have 17 worlds on startup with our data.yml file being 2.8mb. In the old, synchronous behavior this made the World Loading time of our server startup take about 18 seconds. With this change being asynchronous, it took 3s in total.

Further work could be done to retain the loaded config file since it is unlikely to be different on startup. This way you wouldn't be loading the same FileConfiguration X times over. But this seems to work for now I believe.